requirements.txt
discord.py
soundcard
numpy
